atb-gateway-py/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pyjwt==2.8.0
cryptography==42.0.8
requests==2.32.3
